# Improve packet pack performance