--- a/publish.go
+++ b/publish.go
@@ -68,7 +68,7 @@ func (p *pktPublish) Parse(flag byte, contents []byte) (*pktPublish, error) {
 	return p, nil
 }
 
-func (p *pktPublish) pack() []byte {
+func (p *pktPublish) Pack() []byte {
 	pktHeader := packetPublish.b()
 
 	if p.Message.Retain {
@@ -137,7 +137,7 @@ func (c *BaseClient) Publish(ctx context.Context, message *Message) error {
 		c.sig.mu.Unlock()
 	}
 
-	pkt := (&pktPublish{Message: message}).pack()
+	pkt := (&pktPublish{Message: message}).Pack()
 	if err := c.write(pkt); err != nil {
 		return wrapError(err, "sending PUBLISH")
 	}
@@ -158,7 +158,7 @@ func (c *BaseClient) Publish(ctx context.Context, message *Message) error {
 			return ctx.Err()
 		case <-chPubRec:
 		}
-		pktPubRel := (&pktPubRel{ID: message.ID}).pack()
+		pktPubRel := (&pktPubRel{ID: message.ID}).Pack()
 		if err := c.write(pktPubRel); err != nil {
 			return wrapError(err, "sending PUBREL")
 		}