--- a/subscribe.go
+++ b/subscribe.go
@@ -35,7 +35,7 @@ type pktSubscribe struct {
 	Subscriptions []Subscription
 }
 
-func (p *pktSubscribe) pack() []byte {
+func (p *pktSubscribe) Pack() []byte {
 	var payload []byte
 	for _, sub := range p.Subscriptions {
 		payload = append(payload, packString(sub.Topic)...)
@@ -75,7 +75,7 @@ func (c *BaseClient) Subscribe(ctx context.Context, subs ...Subscription) error 
 	c.sig.chSubAck[id] = chSubAck
 	c.sig.mu.Unlock()
 
-	pkt := (&pktSubscribe{ID: id, Subscriptions: subs}).pack()
+	pkt := (&pktSubscribe{ID: id, Subscriptions: subs}).Pack()
 	if err := c.write(pkt); err != nil {
 		return wrapError(err, "sending SUBSCRIBE")
 	}