--- a/unsubscribe.go
+++ b/unsubscribe.go
@@ -23,7 +23,7 @@ type pktUnsubscribe struct {
 	Topics []string
 }
 
-func (p *pktUnsubscribe) pack() []byte {
+func (p *pktUnsubscribe) Pack() []byte {
 	var payload []byte
 	for _, sub := range p.Topics {
 		payload = append(payload, packString(sub)...)
@@ -51,7 +51,7 @@ func (c *BaseClient) Unsubscribe(ctx context.Context, subs ...string) error {
 	c.sig.chUnsubAck[id] = chUnsubAck
 	c.sig.mu.Unlock()
 
-	pkt := (&pktUnsubscribe{ID: id, Topics: subs}).pack()
+	pkt := (&pktUnsubscribe{ID: id, Topics: subs}).Pack()
 	if err := c.write(pkt); err != nil {
 		return wrapError(err, "sending UNSUBSCRIBE")
 	}