master
```
BenchmarkPack/Connect-4           	 2547727	       960 ns/op	     560 B/op	      12 allocs/op
BenchmarkPack/PubAck-4            	18677226	       119 ns/op	      16 B/op	       3 allocs/op
BenchmarkPack/PubComp-4           	18505279	       118 ns/op	      16 B/op	       3 allocs/op
BenchmarkPack/Publish-4           	 6111358	       389 ns/op	     184 B/op	       6 allocs/op
BenchmarkPack/PubRec-4            	18211357	       122 ns/op	      16 B/op	       3 allocs/op
BenchmarkPack/PubRel-4 			18601912	       128 ns/op	      16 B/op	       3 allocs/op
BenchmarkPack/Subscribe-4         	 6444253	       328 ns/op	      48 B/op	       6 allocs/op
BenchmarkPack/Unsubscribe-4       	 7415580	       307 ns/op	      48 B/op	       6 allocs/op
```
this
```
BenchmarkPack/Connect-4           	 5289639	       450 ns/op	     208 B/op       5 allocs/op
BenchmarkPack/PubAck-4            	18627781	       122 ns/op	      16 B/op       3 allocs/op
BenchmarkPack/PubComp-4           	19818740	       121 ns/op	      16 B/op       3 allocs/op
BenchmarkPack/Publish-4 		 8932137	       330 ns/op	     176 B/op	       5 allocs/op
BenchmarkPack/PubRec-4  		19060574	       120 ns/op	      16 B/op	       3 allocs/op
BenchmarkPack/PubRel-4  		17996264	       122 ns/op	      16 B/op	       3 allocs/op
BenchmarkPack/Subscribe-4         	14115542	       161 ns/op	      16 B/op       3 allocs/op
BenchmarkPack/Unsubscribe-4       	15310026	       158 ns/op	      16 B/op       3 allocs/op
```